Represented min_init_stake as uint64

+++ smartcontract/service/native/governance/utils.go
@@ -349,7 +349,7 @@ func CheckVBFTConfig(configuration *config.VBFTConfig) error {
 	if configuration.PeerHandshakeTimeout < 10 {
 		return fmt.Errorf("initConfig. PeerHandshakeTimeout must >= 10")
 	}
-	if configuration.MinInitStake < 10000 {
+	if configuration.MinInitStake < uint64(10000) {
 		return fmt.Errorf("initConfig. MinInitStake must >= 10000")
 	}
 	if len(configuration.VrfProof) < 128 {
